SportsStore/Controllers/AdminController.cs
using System;
using System.Linq;
using Microsoft.AspNetCore.Mvc;
using SportsStore.Models;

namespace SportsStore.Controllers
{
    public class AdminController:Controller
    {
        private readonly IProductRepository _productRepository;

        public AdminController(IProductRepository productRepository)
        {
            _productRepository = productRepository ?? throw new ArgumentNullException(nameof(productRepository));
        }

        public ViewResult Index() => View(_productRepository.Products);

        public ViewResult Edit(int productId) =>
            View(_productRepository.Products.FirstOrDefault(p => p.ProductId == productId));

        [HttpPost]
        public IActionResult Edit(Product product)
        {
            if (!ModelState.IsValid) return View(product);

            _productRepository.SaveProduct(product);
            TempData["message"] = $"{product.Name} has been saved";
            return RedirectToAction("Index");

        }

        public ViewResult Create() => View("Edit", new Product());

        [HttpPost]
        public IActionResult Delete(int productId)
        {
            var deleteProduct = _productRepository.DeleteProduct(productId);

            if (deleteProduct != null)
                TempData["message"] = $"{deleteProduct.Name} was deleted";

            return RedirectToAction("Index");
        }
    }
}
